+++ R/impute_inla.R
@@ -1,6 +1,7 @@
 #' @rdname impute
 #' @importFrom assertthat assert_that is.count
 #' @importFrom dplyr coalesce
+#' @importFrom INLA inla.hyperpar.sample inla.posterior.sample
 #' @importFrom methods new setMethod
 #' @importFrom purrr map map_dfr map2_dfr pmap_dfr
 #' @importFrom stats plogis qpois rgamma rnorm rpois setNames
@@ -39,34 +40,35 @@ setMethod(
     )
 
     dots <- list(...)
+
+    covariates <- get_covariates(model)
     if (missing(extra)) {
-      extra <- model$.args$data[0, ]
+      extra <- covariates[0, ]
     }
 
     response <- as.character(model$.args$formula)[2]
-    missing_obs <- which(is.na(model$.args$data[, response]))
+    missing_obs <- which(is.na(model$.args$data[[response]]))
     if (length(missing_obs) == 0) {
       return(
         new(
-          "rawImputed", Data = model$.args$data, Response = response,
+          "rawImputed", Data = covariates, Response = response,
           Imputation = matrix(integer(0), ncol = n_imp),
           Minimum = coalesce(dots$minimum, ""), Extra = extra
         )
       )
     }
 
-    missing_obs <- sprintf(paste0("Predictor:%i"), missing_obs)
-
-    assert_that(requireNamespace("INLA", quietly = TRUE))
+    ifelse(is.null(model$model.spde2.blc), "Predictor:%i", "APredictor:%i") |>
+      sprintf(missing_obs) -> missing_obs
     assert_that(requireNamespace("sn", quietly = TRUE))
-    samples <- INLA::inla.posterior.sample(
+    samples <- inla.posterior.sample(
       n = n_imp, result = model, seed = seed, num.threads = num_threads,
       parallel.configs = parallel_configs
     )
     map(samples, "latent") |>
       map(`[`, missing_obs, 1) |>
       setNames(paste0("sim_", seq_len(n_imp))) -> latent
-    INLA::inla.hyperpar.sample(n = n_imp, result = model) |>
+    inla.hyperpar.sample(n = n_imp, result = model) |>
       as.data.frame() -> hyperpar
     if (
       model$.args$family == "zeroinflatednbinomial0" &&
@@ -142,9 +144,8 @@ a reproducible example at https://github.com/inbo/multimput/issues"
     )
 
     new(
-      "rawImputed", Data = model$.args$data, Response = response,
-      Imputation = as.matrix(imputation), Minimum = coalesce(dots$minimum, ""),
-      Extra = extra
+      "rawImputed", Data = covariates, Response = response, Extra = extra,
+      Imputation = as.matrix(imputation), Minimum = coalesce(dots$minimum, "")
     )
   }
 )
@@ -186,9 +187,9 @@ rzinb0 <- function(n, mu, size, prob, tol = 2e-10) {
   if (length(low) == n) {
     return(count)
   }
-  dnbinom(x = 0, mu = mu[-low], size = size) |>
+  dnbinom(x = 0, mu = mu[-low], size = size[-low]) |>
     runif(n = n - length(low), max = 1) |>
-    qnbinom(mu = mu[-low], size = size) -> count[non_zero][-low]
+    qnbinom(mu = mu[-low], size = size[-low]) -> count[non_zero][-low]
   return(count)
 }
 
@@ -229,4 +230,24 @@ rzip0 <- function(n, lambda, prob, tol = 2e-10) {
 #' @importFrom stats rbinom rpois
 rzip1 <- function(n, lambda, prob) {
   rbinom(n = n, size = 1, prob = 1 - prob) * rpois(n, lambda = lambda)
+}
+
+get_covariates <- function(model) {
+  if (is.null(model$model.spde2.blc)) {
+    return(as.data.frame(model$.args$data))
+  }
+  covariates <- names(model$.args$data)
+  covariates <- covariates[
+    !covariates %in% c("spde", as.character(model$.args$formula)[2])
+  ]
+  names(model$summary.spde2.blc) |>
+    sprintf(fmt = "^%s") |>
+    grep(covariates, invert = TRUE) -> relevant
+  model$.args$data[covariates[relevant]] |>
+    as.data.frame() -> covariates
+  covariates <- covariates[
+    is.na(model$.args$data[[names(model$summary.spde2.blc)]]),
+  ]
+  model$.args$data[as.character(model$.args$formula)[2]] |>
+    cbind(covariates)
 }