+++ NEWS.md
@@ -1,3 +1,13 @@
+# multimput 0.2.14
+
+* In case all imputed values are identical, `model_impute()` only runs a single
+  model on one imputation.
+  It reports the mean and standard errors based on the single model as-is.
+* `model_impute()` handles empty data.
+* `model_impute()` can filter the covariates with a user supplied function.
+* `model_impute()` gains a `timeout` argument.
+* Bugfix in generating zero-inflated negative binomial data.
+
 # multimput 0.2.13
 
 * `aggregate_impute()` handles the corner case when `join` results in an empty